functions/cut_video/requirements.txt
firebase_functions~=0.1.0
google-cloud-speech
google-cloud-aiplatform
google-cloud-storage
moviepy